Feat: Set up separate route for viewing .epub files
